# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЭкоПульс API",
    "description": "API административной панели эколога/инспектора «ЭкоПульс» — триаж обращений о состоянии площадок ТКО.\n\n**Base URL:** `/api/v1`\n\n**Авторизация:** для защищённых эндпоинтов — Bearer JWT (`Authorization: Bearer <access_token>` из `POST /auth/login`). Приём фотоотчёта — заголовок `X-Intake-Token`. Отдача фото — публично.\n\n**Конверт ошибок:** `{ \"error\": { \"code\", \"message\", \"details\" } }`.\n\nРазделы, помеченные «вне мобильного API», в клиентскую документацию не входят.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/intake/suggest/address": {
      "get": {
        "tags": [
          "Загрузка фото"
        ],
        "summary": "Suggest Address",
        "description": "ПУБЛИЧНО: подсказки адреса для автозаполнения формы волонтёра.\n\nКлюч DaData остаётся на сервере. Без ключа / при q<3 символов / при сбое\nDaData отдаём пустой список (форма деградирует до ручного ввода).\n\nkind ограничивает уровень подсказок (bounded suggest):\n- \"region\"  — только регионы;\n- \"city\"    — города/населённые пункты (в пределах region, если задан);\n- \"street\"  — улицы/дома (в пределах region [+ city], если заданы);\n- \"full\"/иное — полный адрес (поведение по умолчанию).",
        "operationId": "suggest_address_api_v1_intake_suggest_address_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "full",
              "title": "Kind"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 8,
              "title": "Count"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/form": {
      "post": {
        "tags": [
          "Отправка фотоотчёта"
        ],
        "summary": "Public Form",
        "description": "ПУБЛИЧНО: приём обращения из формы волонтёра (multipart/form-data).\n\nwebsite — honeypot: если заполнен, это бот → отдаём ok, но ничего не создаём.",
        "operationId": "public_form_api_v1_intake_form_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_public_form_api_v1_intake_form_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/max": {
      "post": {
        "tags": [
          "Приём вебхуков (server-to-server)"
        ],
        "summary": "Max Intake",
        "description": "Приём обращения из Макс-бота (multipart/form-data) → Incident(source='max').\n\nВызывается сервером Макс-бота server-to-server. Самозащита — общий секрет\nв заголовке X-Intake-Token (тот же гейт, что у /yandex): токен не задан на\nсервере → 503; не совпал → 403. Адрес из текста разбирается DaData Clean API\n(с эвристическим фолбэком). Возвращает {\"ok\": True, \"incident_id\": ...}.",
        "operationId": "max_intake_api_v1_intake_max_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_max_intake_api_v1_intake_max_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/intake/mark-notified": {
      "post": {
        "tags": [
          "Приём вебхуков (server-to-server)"
        ],
        "summary": "Mark Notified",
        "description": "Помечает инциденты как уведомлённые (notified_at = now). Гейт X-Intake-Token.\n\nИдемпотентно: учитывает только ещё не уведомлённые id. Возвращает {\"marked\": N}.",
        "operationId": "mark_notified_api_v1_intake_mark_notified_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNotified"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkNotifiedResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intake/photo/{incident_id}/{filename}": {
      "get": {
        "tags": [
          "Загрузка фото"
        ],
        "summary": "Get Photo",
        "description": "ПУБЛИЧНО: отдаёт байты фото обращения. Жёсткий анти-traversal.\n\nincident_id обязан быть UUID, filename — `^[0-9]+(_thumb)?\\.jpg$`\n(FULL `{i}.jpg` или THUMB `{i}_thumb.jpg`), итоговый путь обязан остаться\nвнутри каталога incidents.",
        "operationId": "get_photo_api_v1_intake_photo__incident_id___filename__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_max_intake_api_v1_intake_max_post": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "msg_id": {
            "type": "string",
            "title": "Msg Id",
            "default": ""
          },
          "msg_url": {
            "type": "string",
            "title": "Msg Url",
            "default": ""
          },
          "sender_name": {
            "type": "string",
            "title": "Sender Name",
            "default": ""
          },
          "photo_time": {
            "type": "string",
            "title": "Photo Time",
            "default": ""
          },
          "photos": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Photos",
            "default": []
          }
        },
        "type": "object",
        "title": "Body_max_intake_api_v1_intake_max_post"
      },
      "Body_public_form_api_v1_intake_form_post": {
        "properties": {
          "fio": {
            "type": "string",
            "title": "Fio",
            "default": ""
          },
          "full_address": {
            "type": "string",
            "title": "Full Address",
            "default": ""
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": ""
          },
          "city": {
            "type": "string",
            "title": "City",
            "default": ""
          },
          "street": {
            "type": "string",
            "title": "Street",
            "default": ""
          },
          "coords": {
            "type": "string",
            "title": "Coords",
            "default": ""
          },
          "photo_time": {
            "type": "string",
            "title": "Photo Time",
            "default": ""
          },
          "bins": {
            "type": "string",
            "title": "Bins",
            "default": ""
          },
          "website": {
            "type": "string",
            "title": "Website",
            "default": ""
          },
          "photos": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Photos",
            "default": []
          }
        },
        "type": "object",
        "title": "Body_public_form_api_v1_intake_form_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkNotified": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "MarkNotified"
      },
      "MarkNotifiedResult": {
        "properties": {
          "marked": {
            "type": "integer",
            "title": "Marked"
          }
        },
        "type": "object",
        "required": [
          "marked"
        ],
        "title": "MarkNotifiedResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}